revert (#12)
* Add GitHub Actions workflow for upstream push

This workflow automates pushing changes from the forked repository to the upstream repository on the main branch.

* Modify workflow to push to Dao-Loc-branch

Updated the GitHub Actions workflow to push changes to a specific branch in the upstream repository.

* Update database config file path in README

Changed the instructions to reference 'app/auth/connect.php' instead of 'app/core/connect.php' for database configuration.

* Update git push command to use x-access-token

* Rename workflow and update push job details

* Delete main.yml